Fix multiple security and code quality issues
Security fixes:
- Fix buffer overflow vulnerability in extend.c protect() function
  Add length validation before strcpy to prevent stack corruption

- Fix unaligned pointer access issues in proxy.c
  Replace direct pointer casting with memcpy for portability
  Fixes map_fix() and addr_equ() functions

Code quality fixes:
- Fix sign comparison warnings across multiple files
  Add proper type casts for size_t/ssize_t comparisons

- Fix format string warnings in main.c parse_cform()
  Cast char* to unsigned char* for %hhx and %hho format specifiers

- Add explicit fall-through comment in proxy.c switch statement
  Clarifies intentional fall-through behavior for UDP command handling

All fixes verified with compilation using -Wall -Wextra flags.

+++ desync.c
@@ -159,7 +159,7 @@ ssize_t send_fake(int sfd, char *buffer,
             p = 0;
             break;
         }
-        memcpy(p, pkt.data, psz < pos ? psz : pos);
+        memcpy(p, pkt.data, psz < (size_t)pos ? psz : (size_t)pos);
         
         if (setttl(sfd, opt->ttl ? opt->ttl : 8, fa) < 0) {
             break;
